Fix passing enum query parameters
When query parameter which is supposed to be an enum (impact, type, etc.)
is passed as a basic type (usually string), bindings fails because
the enum type is expected. Example:

session.retrieve_list(impact="LOW") -> fail
session.retrieve_list(impact=FlawdbApiV1FlawsListImpact("LOW")) -> ok

To fix this we first check whether the passed query param has the
desired type and if not (probably it is a basic type) we pass it in
the expected enum type and after that we can use it.