Replace httpx Client with basic auth storing
Previously we modified the generated AuthenticatedClient to store
httpx Client (session equivalent) which basically stored everything
needed for the requests making. This decision diverted the approach
from the original Python OpenApi Schema client approach.

We are stepping back right now and we will use the AuthenticatedClient
as originally inteded as nothing more than a structure for holding the
neccessary request data and not the httpx session itself.

The only change is that we will store also the auth method so it can
be easily altered.