Replace httpx with requests
Overall the requests library is just much more mature and secure and
recent dependabot scans revealed some very serious issues with the
httpx library, so we are shifting from it to the requests.

Note: httpx will stay only in devel-requirements as it is a dependency
of the openapi-python-client package which is used to build the
bindings. They only use the httpx in some test cases and for demo
purposes in its repository so nothing that should really bother us as
the httpx is essentially not used in our case.