Fix typing get_args usage
After openapi-python-client update to 0.22.0 it started using
native python type annotations (eg. list["AffectPost"] instead of
List[AffectPost]) and thus get_args now only returned string with
name of the inner models instead of the type itself so we now
get just the name and we get the type from the models module.